requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.6.0
certifi==2017.11.5
cffi==1.11.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.4
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.5
django-cors-headers==2.1.0
django-rest-swagger==2.1.2
django-suit==0.2.25
django-toolbelt==0.0.1
djangorestframework==3.7.7
enum34==1.1.6
gunicorn==19.7.1
idna==2.6
ipaddress==1.0.19
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
psycopg2==2.7.3.1
pycparser==2.18
PyJWT==1.5.3
pyOpenSSL==17.5.0
pytz==2017.2
requests==2.18.4
simplejson==3.13.2
six==1.11.0
static3==0.7.0
twilio==6.10.1
uritemplate==3.0.0
urllib3==1.22
